+++ CHANGELOG.md
@@ -5,12 +5,14 @@
 ## [Unreleased]
 
 ### Added
+- approve, unapprove & merge for GitHub
 
 ### Changed
 
 ### Deprecated
 
 ### Removed
+- shallow clone feature, as it messes up the forked repo remotes integration
 
 ### Fixed